Changes to support single node
When parsing the default sqconfig file included in the build,
node name was not getting set because the default sqconfig file
doesn't contain node names, only "virtualnodes".  Added code to
set the node name to a single node in this case.

Added a check for a possible problem with sqgen on a single node
install.  A workaround message is displayed if needed.

Change-Id: I82b7bce943b47c0c879d221c3f9379a75a59f109