Fix for bug 1350641, added log files, other fixes.
Fix for bug 1350641. Now getting cloudera-manager-installer.bin script
from the cloudera website instead of a local HP site. This new version
of cloudera also needed a new java version, so added the download for
that as well.

Added wrapper script for mods scripts 'trafodion_mods' that will
determine which mods script needs to be ran. Wrapper script was added to
make logging easier.

mods, and setup scripts now have a log file that will be located in the
installer logs directory.

Change-Id: Iaa047b456fbf198f8f0161bb2456c200e35466a6