Fix for Bug #1367567 java version detection
A specific java path with a dash in it (/usr/lib/jvm/java-1.6.0)
did not work with the previous method of parsing the java version.
Instead of trying to parse out the java version sub-string,
changed code to look for 1.7. or 1.6.

Also, moved java version detection logic up to trafdoion_mods so its
only done in one location.

Change-Id: Ia2e63d1c7f07e128b2907110d445eded65ae74ef