Merge the private shared single script repo into master.
This is a checkin of the single script project which basically
wraps all the various scripts into one single script that the user
runs.  The script now has the logic to determine which Hadoop distro
is installed as well as to determine if it is a full install or only
and upgrade.

This was developed and tested in a private shared repo so
a full review of the code has already been done.

Change-Id: I45625890f6195f72d56aff77bbafb7a7b84e5895