Merge the private shared single script repo into master.
This is a checkin of the single script project which basically
wraps all the various scripts into one single script that the user
runs.  The script now has the logic to determine which Hadoop distro
is installed as well as to determine if it is a full install or only
an upgrade and many other changes.

This was developed and tested in a private shared repo so
a review of the code has already been done.  QA has approved this
to be checked in.

Change-Id: I37a331c89c311feee430541b69f5f7dfe2177e60