NODE_LIST, removed hostname checks, scanner
Changed how NODE_LIST is created. A list of nodes is entered by the user
and that list is used to ssh to each node given which then the
"hostname" command is used to get hostname which is used to create
NODE_LIST that is used to generate sqconfig. Users can now enter in IP
addresses or node names. All hostname checks are now removed.

The scanner can now be called from trafodion installer.

Change-Id: I5671be3566c1fc3af0d7441254d9fa52150db2ba