Adding Hadoop upgrade support
This code is dependent on a file being added to Trafodion core to work
for Cloudera 5.3 and Hortonworks 2.2. Until that file is added, install
will fail on Cloudera 5.3 and Hortonworks 2.2 but should build fine on
any other version.

Edit 1: Adding check to make nothing greater than Cloudera 5.3.* is
installed or Hortonworks 2.2.* is installed. Adding fix to NFS check.

Edit 2: Copy-paste error.
Change-Id: I065bef3d52651b8cda18deacd0bec6d2343ad1d8